requirements.txt
httpx==0.28.1
pandas==2.2.3
fastapi==0.115.11
uvicorn==0.34.0
python-dotenv==1.0.1
Flask-Mailing==0.2.3
Flask-WTF==1.2.1
Flask==3.1.0
python-dotenv==1.0.1
requests==2.32.3
python-barcode==0.15.1
zeep==4.3.1
Flask-Cors==5.0.0